+++ backend/drizzle/migrations/0003_yummy_senator_kelly.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `curriculum` ADD `scripture2` text;--> statement-breakpoint
+ALTER TABLE `curriculum` ADD `verse_text2` text;